# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Viral Waitlist & Referral Marketing Platform API
  description: |
    REST API for the Viral Waitlist & Referral Marketing Platform.

    ## Authentication

    All API requests require authentication via:
    - **API Key:** Header `X-API-Key: sk_live_xxxxx` or `Authorization: Bearer sk_live_xxxxx`
    - **JWT Token:** For dashboard users, `Authorization: Bearer <jwt>`

    API Key Format:
    - Test: `sk_test_` (test mode)
    - Live: `sk_live_` (production)

    ## Rate Limiting

    Rate limits by tier (per minute):
    - **Free:** 60 requests
    - **Starter:** 60 requests
    - **Pro:** 180 requests
    - **Enterprise:** 600+ requests (custom)

    Response headers include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    - `X-Request-ID`: Unique request identifier for debugging

    ## Error Handling

    All error responses follow a consistent format with both a human-readable message and a machine-readable error code:

    ```json
    {
      "error": "User-friendly error message",
      "code": "ERROR_CODE"
    }
    ```

    Error codes use `UPPER_SNAKE_CASE` format for easy programmatic handling.

    ### Common Error Codes

    - `INVALID_INPUT` - Validation error or malformed request
    - `UNAUTHORIZED` - Missing or invalid authentication
    - `FORBIDDEN` - Insufficient permissions
...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/campaigns/{campaign_id}/users:
    parameters:
      - $ref: '#/components/parameters/CampaignIdParam'

    post:
      tags:
        - Waitlist Users
      summary: Sign up a user to the waitlist (Public)
      description: |
        Add a new user to the campaign waitlist. This is a **public endpoint** that does not require authentication.

        Only `email` and `terms_accepted` are required fields. Additional user information can be provided via `custom_fields`.
      operationId: signupUser
      security: []  # Public endpoint - no authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            examples:
              minimal:
                summary: Minimal signup (required fields only)
                value:
                  email: "user@example.com"
                  terms_accepted: true
              with_custom_fields:
                summary: Signup with custom fields
                value:
                  email: "user@example.com"
                  terms_accepted: true
                  custom_fields:
                    first_name: "John"
                    last_name: "Doe"
                    company: "Acme Inc"
              with_referral:
                summary: Signup with referral code
                value:
                  email: "user@example.com"
                  terms_accepted: true
                  referral_code: "ABC123XY"
                  custom_fields:
                    first_name: "Jane"
                    last_name: "Smith"
              with_utm:
                summary: Signup with UTM tracking
                value:
                  email: "user@example.com"
                  terms_accepted: true
                  utm_source: "facebook"
                  utm_medium: "social"
                  utm_campaign: "spring-launch"
                  marketing_consent: true
      responses:
        '201':
          description: User signed up successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Campaign not found"
                code: "CAMPAIGN_NOT_FOUND"
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Waitlist Users
      summary: List waitlist users
      description: Retrieve a paginated list of waitlist users for a campaign
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by user status
          schema:
            type: string
            enum: [pending, verified, converted, removed, blocked]
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
            enum: [position, created_at, referral_count]
            default: position
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of waitlist users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}/users/{user_id}:
    parameters:
      - $ref: '#/components/parameters/CampaignIdParam'
      - $ref: '#/components/parameters/UserIdParam'

    get:
      tags:
        - Waitlist Users
      summary: Get user details
      description: Retrieve detailed information about a specific waitlist user
      operationId: getUser
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Waitlist Users
      summary: Update user information
      description: Update a waitlist user's information
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Waitlist Users
      summary: Remove user from waitlist
      description: Remove a user from the waitlist (soft delete)
      operationId: deleteUser
      responses:
        '204':
          description: User removed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/campaigns/{campaign_id}/users/search:
    post:
      tags:
        - Waitlist Users
      summary: Search waitlist users
      description: Search and filter waitlist users with advanced criteria
      operationId: searchUsers
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Search query (email, name)
                status:
                  type: array
                  items:
                    type: string
                    enum: [pending, verified, converted, removed, blocked]
                verified:
                  type: boolean
                  description: Filter by email verification status
                min_referrals:
                  type: integer
                  description: Minimum referral count
                date_from:
                  type: string
                  format: date-time
                  description: Filter users created after this date
                date_to:
                  type: string
                  format: date-time
                  description: Filter users created before this date
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/campaigns/{campaign_id}/users/{user_id}/verify:
    post:
      tags:
        - Waitlist Users
      summary: Verify user email
      description: Manually verify a user's email address
      operationId: verifyUser
      parameters:
        - $ref: '#/components/parameters/CampaignIdParam'
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication (sk_live_* or sk_test_*)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for dashboard users
  parameters:
    CampaignIdParam:
      name: campaign_id
      in: path
      required: true
      description: Campaign unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              value:
                error: "Invalid request format. Please check your JSON syntax."
                code: "INVALID_INPUT"
            missing_field:
              value:
                error: "Required field is missing"
                code: "INVALID_INPUT"
...
    Conflict:
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            email_exists:
              value:
                error: "Email already exists"
                code: "EMAIL_EXISTS"
            user_exists:
              value:
                error: "User already exists in this campaign"
                code: "USER_EXISTS"
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum requests per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Please try again later."
            code: "RATE_LIMIT_EXCEEDED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "An internal server error occurred. Please try again later."
            code: "INTERNAL_ERROR"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: User-friendly error message
        code:
          type: string
          description: Machine-readable error code in UPPER_SNAKE_CASE format
          example: "INVALID_INPUT"
      example:
        error: "Invalid request format"
        code: "INVALID_INPUT"

    # ==================== CORE MODELS ====================
...
    SignupRequest:
      type: object
      required:
        - email
        - terms_accepted
      properties:
        email:
          type: string
          format: email
          description: User's email address (required)
        referral_code:
          type: string
          description: Referral code if referred by another user
        custom_fields:
          type: object
          additionalProperties:
            type: string
          description: Custom form fields (e.g., first_name, last_name, company, etc.)
        utm_source:
          type: string
          description: UTM source parameter for tracking
        utm_medium:
          type: string
          description: UTM medium parameter for tracking
        utm_campaign:
          type: string
          description: UTM campaign parameter for tracking
        utm_term:
          type: string
          description: UTM term parameter for tracking
        utm_content:
          type: string
          description: UTM content parameter for tracking
        marketing_consent:
          type: boolean
          default: false
          description: Whether user consents to marketing communications
        terms_accepted:
          type: boolean
          description: Whether user accepted terms and conditions (required)
...
    SignupResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/WaitlistUser'
        position:
          type: integer
        referral_link:
          type: string
          format: uri
        message:
          type: string
...